api/v1/orders/serializers.py
from rest_framework import serializers

from apps.order.models import OrderItem


class AddOrderItemQuerySerializer(serializers.Serializer):
    product_id = serializers.IntegerField()
    quantity = serializers.IntegerField(min_value=1)


class OrderItemSerializer(serializers.ModelSerializer):

    class Meta:
        model = OrderItem
        fields = '__all__'
