requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
dotenv==0.9.9
pi==0.1.2
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
